requirements.txt
anyio==3.6.2
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
fastapi==0.88.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
Mastodon.py==1.8.0
pydantic==1.10.2
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0